Split delayed into separate grafana metrics and some refactoring
